Fix tying
