+++ default/scripting/focs/_effects.pyi
@@ -16,6 +16,7 @@ from focs._types import (
     _Effect,
     _EffectGroup,
     _Empire,
+    _EmpireId,
     _FloatParam,
     _Focus,
     _IntParam,
@@ -212,15 +213,15 @@ def NamedReal(*, name: str, value: _FloatParam) -> float: ...
 def NamedRealLookup(*, name: str) -> float: ...
 def NamedIntegerLookup(*, name: str) -> int: ...
 def JumpsBetween(source: _ID, target: _ID) -> int: ...
-def SpeciesShipsDestroyed(empire: _Empire, name: _SpeciesValue) -> int: ...
-def SpeciesShipsLost(empire: _Empire, name: _SpeciesValue) -> int: ...
-def SpeciesEmpireTargetOpinion(*, species: _SpeciesValue, empire: _Empire) -> float: ...
-def EmpireStockpile(empire: _Empire, resource: _Resource) -> float: ...
+def SpeciesShipsDestroyed(empire: _EmpireId, name: _SpeciesValue) -> int: ...
+def SpeciesShipsLost(empire: _EmpireId, name: _SpeciesValue) -> int: ...
+def SpeciesEmpireTargetOpinion(*, species: _SpeciesValue, empire: _EmpireId) -> float: ...
+def EmpireStockpile(empire: _EmpireId, resource: _Resource) -> float: ...
 def PartCapacity(*, name: str) -> float: ...
 def PartSecondaryStat(*, name: str) -> float: ...
 def PartsInShipDesign(*, name: str, design: _DesignID) -> float: ...
 def UserString(name: str) -> str: ...
-def EmpireMeterValue(empire: _Empire, meter: str) -> float: ...
+def EmpireMeterValue(empire: _EmpireId, meter: str) -> float: ...
 def EffectsGroup(
     *,
     scope: _Condition,
@@ -237,7 +238,7 @@ def EffectsGroup(
 CurrentContent = ""
 
 def Enqueued(type: _BuildingType, name: str) -> _Condition: ...
-def TurnTechResearched(*, empire: _Empire, name: str) -> _Condition: ...
+def TurnTechResearched(*, empire: _EmpireId, name: str) -> _Condition: ...
 def IsBuilding(*, name: list[str | _BuildingType] = ...) -> _Condition: ...
 def Number(*, low: int, condition: _Condition) -> _Condition: ...
 def Planet(
@@ -253,7 +254,7 @@ def SpeciesLikes(*, name: _Focus) -> _Condition: ...
 def SpeciesDislikes(*, name: _Focus) -> _Condition: ...
 def Homeworld(*, name: list[_SpeciesValue] = ...) -> _Condition: ...
 def HasTag(*, name: str) -> _Condition: ...
-def VisibleToEmpire(*, empire: _Empire) -> _Condition: ...
+def VisibleToEmpire(*, empire: _EmpireId) -> _Condition: ...
 def ContainedBy(condition: _Condition) -> _Condition: ...
 def HasSpecial(*, name: str) -> _Condition: ...
 def OwnerHasTech(*, name: str) -> _Condition: ...
@@ -281,21 +282,21 @@ def TargetIndustry(*, low: int) -> _Condition: ...
 def Happiness(*, low: _FloatParam) -> _Condition: ...
 def Focus(*, type: list[str]) -> _Condition: ...
 def Random(*, probability: _FloatParam) -> _Condition: ...
-def OwnedBy(*, affiliation: _Empire = _Empire(), empire: _Empire = _Empire()) -> _Condition: ...
+def OwnedBy(*, affiliation: _EmpireId = _Empire(), empire: _EmpireId = _Empire()) -> _Condition: ...
 def WithinStarlaneJumps(*, jumps: _IntParam, condition: _Condition) -> _Condition: ...
 def WithinDistance(*, distance: int, condition: _Condition) -> _Condition: ...
-def ResourceSupplyConnected(*, empire: _Empire, condition: _Condition) -> _Condition: ...
+def ResourceSupplyConnected(*, empire: _EmpireId, condition: _Condition) -> _Condition: ...
 def Contains(scope: _Condition) -> _Condition: ...
 def Turn(*, high: _IntParam = ..., low: _IntParam = ...) -> _Condition: ...
 def Structure(*, low: _FloatParam = ..., high: _FloatParam = ...) -> _Condition: ...
-def ResupplyableBy(*, empire: _Empire) -> _Condition: ...
+def ResupplyableBy(*, empire: _EmpireId) -> _Condition: ...
 def DesignHasPart(*, name: str, low: int = ..., high: int = ...) -> _Condition: ...
-def EmpireHasAdoptedPolicy(*, name: str, empire: _Empire = _Empire()) -> _Condition: ...
+def EmpireHasAdoptedPolicy(*, name: str, empire: _EmpireId = _Empire()) -> _Condition: ...
 
 Influence = _Resource()
 
-def HasEmpireStockpile(*, empire: _Empire, resource: _Resource, low: int = ..., high: int = ...) -> _Condition: ...
-def NumPoliciesAdopted(*, empire: _Empire) -> _Condition: ...
+def HasEmpireStockpile(*, empire: _EmpireId, resource: _Resource, low: int = ..., high: int = ...) -> _Condition: ...
+def NumPoliciesAdopted(*, empire: _EmpireId) -> _Condition: ...
 def OnPlanet(*, id: _ID) -> _Condition: ...
 
 # Effects
@@ -332,8 +333,8 @@ def SetMaxTroops(*, value: _FloatParam) -> _Effect: ...
 def SetStockpile(*, value: _FloatParam) -> _Effect: ...
 def SetMaxStockpile(*, value: _FloatParam, accountinglabel: str = "") -> _Effect: ...
 def SetTargetHappiness(*, value: _FloatParam) -> _Effect: ...
-def GiveEmpirePolicy(*, name: str, empire: _Empire = _Empire()) -> _Effect: ...
-def SetVisibility(empire: _Empire, visibility: Any) -> _Effect: ...
+def GiveEmpirePolicy(*, name: str, empire: _EmpireId = _Empire()) -> _Effect: ...
+def SetVisibility(empire: _EmpireId, visibility: Any) -> _Effect: ...
 def SetPopulation(*, value: _FloatParam) -> _Effect: ...
 def SetTroops(*, value: _FloatParam) -> _Effect: ...
 def SetMaxDefense(*, value: _FloatParam) -> _Effect: ...
@@ -346,8 +347,8 @@ def SetMaxFuel(*, value: _FloatParam) -> _Effect: ...
 def SetResearch(*, value: _FloatParam) -> _Effect: ...
 def SetIndustry(*, value: _FloatParam) -> _Effect: ...
 def SetInfluence(*, value: _FloatParam) -> _Effect: ...
-def SetSpeciesTargetOpinion(*, species: _SpeciesValue, empire: _Empire, opinion: _FloatParam) -> _Effect: ...
-def SetSpeciesOpinion(*, species: _SpeciesValue, empire: _Empire, opinion: _FloatParam) -> _Effect: ...
+def SetSpeciesTargetOpinion(*, species: _SpeciesValue, empire: _EmpireId, opinion: _FloatParam) -> _Effect: ...
+def SetSpeciesOpinion(*, species: _SpeciesValue, empire: _EmpireId, opinion: _FloatParam) -> _Effect: ...
 def SetTargetPopulation(*, value: _FloatParam, accountinglabel: str = "") -> _Effect: ...
 def SetTargetInfluence(*, value: _FloatParam) -> _Effect: ...
 def SetTargetConstruction(*, value: _FloatParam) -> _Effect: ...
@@ -364,15 +365,15 @@ def SetMaxSecondaryStat(*, partname: str, value: _FloatParam) -> _Effect: ...
 def MoveTo(*, destination: _Condition) -> _Effect: ...
 def MoveTowards(*, speed: int, target: _Condition) -> _Effect: ...
 def SetStarType(*, type: _StarType) -> _Effect: ...
-def SetEmpireMeter(*, empire: _Empire, meter: str, value: _FloatParam) -> _Effect: ...
-def SetOwner(*, empire: _Empire) -> _Effect: ...
+def SetEmpireMeter(*, empire: _EmpireId, meter: str, value: _FloatParam) -> _Effect: ...
+def SetOwner(*, empire: _EmpireId) -> _Effect: ...
 def GenerateSitRepMessage(
     *,
     message: str,
     label: str,
-    empire: _Empire,
+    empire: _EmpireId,
     parameters: dict[str, Any] = {},
     icon: str = "",
 ) -> _Effect: ...
-def GiveEmpireTech(*, name: str, empire=_Empire()) -> _Effect: ...
+def GiveEmpireTech(*, name: str, empire: _EmpireId = _Empire()) -> _Effect: ...
 def Victory(*, reason: str) -> _Effect: ...
